rename checks to check
