+++ zebra-consensus/src/block.rs
@@ -8,12 +8,11 @@
 //! Verification is provided via a `tower::Service`, to support backpressure and batch
 //! verification.
 
-mod checks;
+mod check;
 
 #[cfg(test)]
 mod tests;
 
-use checks::*;
 use chrono::Utc;
 use color_eyre::eyre::{eyre, Report};
 use futures_util::FutureExt;
@@ -109,7 +108,7 @@ where
             // Field validity and structure checks
             let now = Utc::now();
             block.header.is_time_valid_at(now)?;
-            is_coinbase_first(&block)?;
+            check::is_coinbase_first(&block)?;
 
             // TODO:
             //   - context-free header verification: merkle root
